+++ src/blocks/features/feature/index.js
@@ -1,21 +1,16 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * Internal dependencies
  */
-import { BackgroundStyles, BackgroundAttributes, BackgroundClasses, BackgroundVideo } from '../../../components/background';
+import { BackgroundAttributes } from '../../../components/background';
 import DimensionsAttributes from '../../../components/dimensions-control/attributes';
-import Edit from './components/edit';
-import icons from './components/icons';
+import edit from './edit';
+import icons from './icons';
+import save from './save';
 
 /**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { InnerBlocks, getColorClassName } = wp.blockEditor;
 
 /**
  * Block constants
@@ -42,7 +37,7 @@ const blockAttributes = {
 
 const settings = {
 
-	title: title,
+	title,
 
 	description: __( 'A singular child column within a parent features block.' ),
 
@@ -54,49 +49,9 @@ const settings = {
 		inserter: false,
 	},
 
-	edit: Edit,
+	edit,
 
-	save( { attributes, className } ) {
-		const {
-			coblocks,
-			contentAlign,
-			customTextColor,
-			textColor,
-			paddingSize,
-		} = attributes;
-
-		// Body color class and styles.
-		const textClass = getColorClassName( 'color', textColor );
-
-		const classes = classnames(
-			className, {
-				[ `has-${ contentAlign }-content` ]: contentAlign,
-				[ `coblocks-feature-${ coblocks.id }` ]: coblocks && ( typeof coblocks.id !== 'undefined' ),
-			} );
-
-		const innerClasses = classnames(
-			'wp-block-coblocks-feature__inner',
-			...BackgroundClasses( attributes ), {
-				'has-text-color': textColor || customTextColor,
-				[ textClass ]: textClass,
-				'has-padding': paddingSize && paddingSize !== 'no',
-				[ `has-${ paddingSize }-padding` ]: paddingSize && ( paddingSize !== 'advanced' ),
-			} );
-
-		const innerStyles = {
-			...BackgroundStyles( attributes ),
-			color: textClass ? undefined : customTextColor,
-		};
-
-		return (
-			<div className={ classes }>
-				<div className={ innerClasses } style={ innerStyles }>
-					{ BackgroundVideo( attributes ) }
-					<InnerBlocks.Content />
-				</div>
-			</div>
-		);
-	},
+	save,
 };
 
 export { name, title, icon, settings };
