+++ src/blocks/features/inspector.js
@@ -2,8 +2,8 @@
  * Internal dependencies
  */
 import applyWithColors from './colors';
-import { BackgroundPanel } from '../../../../components/background';
-import DimensionsControl from '../../../../components/dimensions-control/';
+import { BackgroundPanel } from '../../components/background';
+import DimensionsControl from '../../components/dimensions-control/';
 
 /**
  * WordPress dependencies
@@ -12,7 +12,7 @@ const { __ } = wp.i18n;
 const { Component, Fragment } = wp.element;
 const { compose } = wp.compose;
 const { InspectorControls, PanelColorSettings } = wp.blockEditor;
-const { PanelBody, withFallbackStyles } = wp.components;
+const { PanelBody, RangeControl, SelectControl, withFallbackStyles } = wp.components;
 
 /**
  * Fallback styles
@@ -34,6 +34,7 @@ const applyFallbackStyles = withFallbackStyles( ( node, ownProps ) => {
 class Inspector extends Component {
 	render() {
 		const {
+			clientId,
 			attributes,
 			backgroundColor,
 			setAttributes,
@@ -43,6 +44,25 @@ class Inspector extends Component {
 		} = this.props;
 
 		const {
+			columns,
+			gutter,
+			marginBottom,
+			marginLeft,
+			marginRight,
+			marginSize,
+			marginTop,
+			marginBottomTablet,
+			marginLeftTablet,
+			marginRightTablet,
+			marginTopTablet,
+			marginBottomMobile,
+			marginLeftMobile,
+			marginRightMobile,
+			marginTopMobile,
+			marginSyncUnits,
+			marginSyncUnitsTablet,
+			marginSyncUnitsMobile,
+			marginUnit,
 			paddingBottom,
 			paddingLeft,
 			paddingRight,
@@ -62,10 +82,62 @@ class Inspector extends Component {
 			paddingUnit,
 		} = attributes;
 
+		const gutterOptions = [
+			{ value: 'small', label: __( 'Small' ) },
+			{ value: 'medium', label: __( 'Medium' ) },
+			{ value: 'large', label: __( 'Large' ) },
+		];
+
 		return (
 			<Fragment>
 				<InspectorControls>
-					<PanelBody title={ __( 'Feature Settings' ) } className="components-coblocks-block-sidebar--features">
+					<PanelBody title={ __( 'Features Settings' ) } className="components-coblocks-block-sidebar--features">
+						<RangeControl
+							label={ __( 'Columns' ) }
+							value={ columns }
+							onChange={ ( nextCount ) => {
+								setAttributes( {
+									columns: parseInt( nextCount ),
+								} );
+
+								if ( parseInt( nextCount ) < 2 ) {
+									setAttributes( {
+										gutter: 'no',
+									} );
+								} else {
+									if ( gutter === 'no' ) {
+										setAttributes( {
+											gutter: 'large',
+										} );
+									}
+								}
+
+								wp.data.dispatch( 'core/block-editor' ).selectBlock( clientId );
+							} }
+							min={ 1 }
+							max={ 3 }
+						/>
+						<DimensionsControl { ...this.props }
+							type={ 'margin' }
+							label={ __( 'Margin' ) }
+							valueTop={ marginTop }
+							valueRight={ marginRight }
+							valueBottom={ marginBottom }
+							valueLeft={ marginLeft }
+							valueTopTablet={ marginTopTablet }
+							valueRightTablet={ marginRightTablet }
+							valueBottomTablet={ marginBottomTablet }
+							valueLeftTablet={ marginLeftTablet }
+							valueTopMobile={ marginTopMobile }
+							valueRightMobile={ marginRightMobile }
+							valueBottomMobile={ marginBottomMobile }
+							valueLeftMobile={ marginLeftMobile }
+							unit={ marginUnit }
+							syncUnits={ marginSyncUnits }
+							syncUnitsTablet={ marginSyncUnitsTablet }
+							syncUnitsMobile={ marginSyncUnitsMobile }
+							dimensionSize={ marginSize }
+						/>
 						<DimensionsControl { ...this.props }
 							type={ 'padding' }
 							label={ __( 'Padding' ) }
@@ -87,6 +159,15 @@ class Inspector extends Component {
 							syncUnitsMobile={ paddingSyncUnitsMobile }
 							dimensionSize={ paddingSize }
 						/>
+						{ columns >= 2 &&
+						<SelectControl
+							label={ __( 'Gutter' ) }
+							value={ gutter }
+							options={ gutterOptions }
+							help={ __( 'Space between each column.' ) }
+							onChange={ ( value ) => setAttributes( { gutter: value } ) }
+						/>
+						}
 					</PanelBody>
 					<PanelColorSettings
 						title={ __( 'Color Settings' ) }
@@ -117,7 +198,9 @@ class Inspector extends Component {
 					>
 					</PanelColorSettings>
 					<BackgroundPanel { ...this.props }
+						hasCaption={ false }
 						hasOverlay={ true }
+						hasGalleryControls={ false }
 					/>
 				</InspectorControls>
 			</Fragment>
